+++ src/graphql_clj/validator/rules/scalar_leafs.clj
@@ -5,12 +5,15 @@
             [clojure.spec :as s]
             [graphql-clj.spec :as spec]))
 
-(defn- required-subselection-error [field-name field-type]
-  {:error (format "Field '%s' of type '%s' must have a selection of subfields." field-name (name field-type))
+(defn- field-type+spec->label [field-type spec]
+  (if (keyword? field-type) (name field-type) (name spec)))
+
+(defn- required-subselection-error [field-name field-type spec]
+  {:error (format "Field '%s' of type '%s' must have a selection of subfields." field-name (field-type+spec->label field-type spec))
    :loc   (ve/extract-loc (meta field-name))})
 
-(defn- no-subselection-allowed-error [field-name field-type]
-  {:error (format "Field '%s' must not have a selection since type '%s' has no subfields." field-name (name field-type))
+(defn- no-subselection-allowed-error [field-name field-type spec]
+  {:error (format "Field '%s' must not have a selection since type '%s' has no subfields." field-name (field-type+spec->label field-type spec))
    :loc   (ve/extract-loc (meta field-name))})
 
 (def scalar-kinds #{:SCALAR :ENUM})
@@ -22,11 +25,11 @@
                     (scalar-kinds (:kind (or (spec/get-type-node field-type s)
                                              (spec/get-type-node spec s)))))]
     (cond (and scalar? selection-set)
-          {:state (ve/update-errors s (no-subselection-allowed-error field-name field-type))
+          {:state (ve/update-errors s (no-subselection-allowed-error field-name field-type spec))
            :break true}
 
           (and (not scalar?) (not selection-set))
-          {:state (ve/update-errors s (required-subselection-error field-name field-type))
+          {:state (ve/update-errors s (required-subselection-error field-name field-type spec))
            :break true})))
 
 (def rules [non-scalar-leaf])