+++ src/graphql_clj/validator/rules/unique_fragment_names.clj
@@ -3,7 +3,7 @@
   (:require [graphql-clj.visitor :refer [defnodevisitor]]
             [graphql-clj.validator.errors :as ve]))
 
-(defnodevisitor duplicate-fragment-name :pre :query-root
+(defnodevisitor duplicate-fragment-name :pre :statement-root
   [{:keys [children] :as n} s]
   (ve/guard-duplicate-names "fragment" :name (filter #(= :fragment-definition (:node-type %)) children) s))